Converted to Vue 2.6.1 to use vuetify
