+++ .babelrc
@@ -1,6 +1,5 @@
 {
   "presets": [
-    ["env", { "modules": false }],
-    "stage-3"
+    ["@babel/env", { "modules": false }]
   ]
 }
