+++ README.md
@@ -1,3 +1,24 @@
-# PersonalSiteFrontEnd
-This is the front end of my personal site built with Vue.js
-Current Version visible at https://rybeau.web.app/ NOTE: Backend is not currently deployed so only test information is displayed currently.
+# personal-site
+
+## Project setup
+```
+npm install
+```
+
+### Compiles and hot-reloads for development
+```
+npm run serve
+```
+
+### Compiles and minifies for production
+```
+npm run build
+```
+
+### Lints and fixes files
+```
+npm run lint
+```
+
+### Customize configuration
+See [Configuration Reference](https://cli.vuejs.org/config/).
