+++ package.json
@@ -1,44 +1,58 @@
 {
   "name": "personal-site",
-  "description": "This the front end to my personal site built with vue js",
-  "version": "1.0.0",
-  "author": "Ryan Beaumont",
-  "license": "MIT",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --open --hot",
-    "build": "cross-env NODE_ENV=production webpack --progress --hide-modules"
+    "preinstall": "npx npm-force-resolutions",
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.19.2",
-    "moment": "^2.27.0",
-    "vue": "^2.5.11",
-    "vue-axios": "^2.1.5",
-    "vue-router": "^3.3.4",
-    "vuetify": "^2.3.6",
-    "vuex": "^3.5.1",
-    "webpack-cli": "^3.3.12"
+    "core-js": "^3.6.5",
+    "vue": "^2.6.11",
+    "vue-router": "^3.2.0",
+    "vuetify": "^2.2.11",
+    "vuex": "^3.4.0",
+    "moment": "2.27.0"
+  },
+  "devDependencies": {
+    "@vue/cli-plugin-babel": "~4.4.0",
+    "@vue/cli-plugin-eslint": "~4.4.0",
+    "@vue/cli-plugin-router": "~4.4.0",
+    "@vue/cli-plugin-vuex": "~4.4.0",
+    "@vue/cli-service": "~4.4.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.7.2",
+    "eslint-plugin-vue": "^6.2.2",
+    "sass": "^1.19.0",
+    "sass-loader": "^8.0.0",
+    "vue-cli-plugin-vuetify": "~2.0.7",
+    "vue-template-compiler": "^2.6.11",
+    "vuetify-loader": "^1.3.0"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "babel-eslint"
+    },
+    "rules": {
+      "no-unused-vars": "off"
+    }
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not dead"
   ],
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-stage-3": "^6.24.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^3.6.0",
-    "deepmerge": "^4.2.2",
-    "fibers": "^5.0.0",
-    "file-loader": "^1.1.4",
-    "sass": "^1.26.10",
-    "sass-loader": "^9.0.2",
-    "vue-loader": "^15.9.3",
-    "vue-template-compiler": "^2.4.4",
-    "webpack": "^4.43.0",
-    "webpack-dev-server": "^3.11.0"
+  "resolutions": {
+    "serialize-javascript": "^3.1.0"
   }
 }
