+++ src/plugins/vuetify.js
@@ -1,9 +1,7 @@
-import Vue from 'vue'
-import Vuetify from 'vuetify'
-import 'vuetify/dist/vuetify.min.css'
+import Vue from 'vue';
+import Vuetify from 'vuetify/lib';
 
 Vue.use(Vuetify);
 
-const opts = {};
-
-export default new Vuetify(opts);
+export default new Vuetify({
+});
