+++ src/router/index.js
@@ -1,0 +1,76 @@
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+import Home from '../views/Home.vue'
+import Blog from '../views/Blog.vue'
+import Contact from '../views/Contact.vue'
+import Projects from '../views/Projects.vue'
+import Profile from '../views/Profile.vue'
+import Error from '../views/Error.vue'
+
+Vue.use(VueRouter)
+
+const routes = [
+  {
+    path: '/',
+    name: 'Home',
+    component: Home,
+    meta: {
+      title: 'Home - Ryan Beaumont'
+    }
+  },
+  {
+    path: '/blog',
+    name: 'Blog',
+    component: Blog,
+    meta: {
+      title: 'Blog - Ryan Beaumont'
+    }
+  },
+  {
+    path: '/projects',
+    name: 'Projects',
+    component: Projects,
+    meta: {
+      title: 'Projects - Ryan Beaumont'
+    }
+  },
+  {
+    path: '/contact',
+    name: 'Contact',
+    component: Contact,
+    meta: {
+      title: 'Contact - Ryan Beaumont'
+    }
+  },
+  {
+    path: '/profile',
+    name: 'Profile',
+    component: Profile,
+    meta: {
+      title: 'Your Profile'
+    }
+  },
+  {
+    path: '*',
+    name: 'Error',
+    component: Error,
+    meta: {
+      title: 'Oops - Error'
+    }
+  }
+
+];
+
+const router = new VueRouter({
+  mode: 'history',
+  base: process.env.BASE_URL,
+  routes
+})
+
+router.afterEach(() => {
+  Vue.nextTick((to, _) => {
+    document.title = to.meta.title;
+  });
+});
+
+export default router
